requirements.txt
# Cloud Provider SDKs
boto3>=1.28.0
azure-mgmt-compute>=30.0.0
azure-mgmt-storage>=21.0.0
azure-mgmt-sql>=4.0.0
azure-identity>=1.14.0
google-cloud-compute>=1.14.0
google-cloud-storage>=2.10.0
google-cloud-sql>=1.6.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Configuration
PyYAML>=6.0

# CLI
click>=8.1.0
rich>=13.5.0

# Reporting
jinja2>=3.1.0
matplotlib>=3.7.0
plotly>=5.16.0

# AI/ML (optional)
openai>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code Quality
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
